+++ src/Services/QueryUserInformationService.ts
@@ -1,9 +1,34 @@
 import axios from "axios"
 import { baseUrls, getHeaders } from "./ServiceHelpers"
-import { QueryUserInformationService } from "./Services"
 import { UserInfo, Vehicles } from "./ServiceTypes"
 
 const baseUrl = baseUrls.IF9_BASE_URL
+
+/**
+ * Queries User Information
+ */
+export interface QueryUserInformationService {
+    /**
+     * List user information based on the user ID.
+     * 
+     * @param accessToken Access Token
+     * @param deviceId UUID4 Device Identifier
+     * @param userId User Identifier
+     */
+    getUserInfoFromId: (accessToken: string, deviceId: string, userId: string) => Promise<UserInfo>
+
+    /**
+     * Lists the vehicles associated with the specified user id. It is unclear at this time what
+     * the `primaryOnly` parameter does as all vehicles associated with the user id will be
+     * returned regardless. It is possible that this parameter is simply not used for anything.
+     * 
+     * @param accessToken Access Token
+     * @param deviceId UUID4 Device Identifier
+     * @param userId User Identifier
+     * @param primaryOnly Boolean value indicating whether to query the primary vehicle only
+     */
+    getVehiclesForUser: (accessToken: string, deviceId: string, userId: string, primaryOnly?: boolean) => Promise<Vehicles>
+}
 
 const queryUserInformationService: QueryUserInformationService = {
     getUserInfoFromId: async (accessToken: string, deviceId: string, userId: string): Promise<UserInfo> => {
