+++ src/Services/ServiceHelpers.ts
@@ -1,11 +1,7 @@
-const IFAS_BASE_URL = 'https://ifas.prod-row.jlrmotor.com/ifas/jlr'
-const IFOP_BASE_ULR = 'https://ifop.prod-row.jlrmotor.com/ifop/jlr'
-const IF9_BASE_URL = 'https://if9.prod-row.jlrmotor.com/if9/jlr'
-
 const baseUrls = {
-    IFAS_BASE_URL,
-    IFOP_BASE_ULR,
-    IF9_BASE_URL
+    IFAS_BASE_URL: 'https://ifas.prod-row.jlrmotor.com/ifas/jlr',
+    IFOP_BASE_URL: 'https://ifop.prod-row.jlrmotor.com/ifop/jlr',
+    IF9_BASE_URL:  'https://if9.prod-row.jlrmotor.com/if9/jlr'
 }
 
 const getHeaders = (accessToken: string, deviceId: string, partial?: any) => {
@@ -17,4 +13,4 @@ const getHeaders = (accessToken: string, deviceId: string, partial?: any) => {
     }
 }
 
-export {getHeaders, baseUrls}
+export { getHeaders, baseUrls }
